fix and and add "print reg."

+++ src/kernel/inc/types.h
@@ -6,7 +6,7 @@ typedef unsigned short uint16_t;
 typedef unsigned int  uint32_t;
 typedef unsigned long long uint64_t;
 
-typedef uint64_t ssize_t;
-typedef ssize_t reg_t;
+typedef long long ssize_t;
+typedef uint64_t reg_t;
 
 #endif /* __TYPES_H__ */
